# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/internal_api.UserResponse'
    type: object
...
  internal_api.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: bad request
        type: string
      message:
        example: Invalid input
        type: string
    type: object
...
  internal_api.LoginRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  internal_api.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  internal_api.UserResponse:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      display_name:
        example: John Doe
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: BeeCount Cloud API 是一个用于管理个人财务账本的云服务 API，提供账本、交易、分类、标签等功能。
  license:
    name: AGPL 3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: http://swagger.io/terms/
  title: BeeCount Cloud API
  version: "1.0"
paths:
  /api/v1/attachments/{id}:
    delete:
      consumes:
      - application/json
      description: 从存储后端删除文件，并删除数据库中的附件记录
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 删除成功
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除附件
      tags:
      - 附件
    get:
      consumes:
      - application/json
      description: 根据附件ID下载文件并返回给客户端
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取附件
      tags:
      - 附件
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户凭证，返回JWT token和用户信息
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/internal_api.AuthResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: 注册已禁用
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: 用户注册
      tags:
      - 认证
...
  /api/v1/ledgers/{ledger_id}/transactions/{transaction_id}/attachments:
    post:
      consumes:
      - multipart/form-data
      description: 上传交易附件，支持图片格式（JPEG、PNG、GIF、WebP），最大10MB
      parameters:
      - description: 账本ID
        in: path
        name: ledger_id
        required: true
        type: integer
      - description: 交易ID
        in: path
        name: transaction_id
        required: true
        type: integer
      - description: 附件文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: 上传成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 上传附件
      tags:
      - 附件
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: 根据JWT token返回当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取当前用户信息
      tags:
      - 认证
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"